requirements.txt
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
PIMS==0.6.1
pytorch_lightning==2.0
pytorchvideo==0.1.5
PyYAML==6.0.1
scikit_learn==1.0.2
scipy==1.7.3
torch==2.0.1
torchbnn==1.2
torchvision==0.15.2
tqdm==4.57.0
tsai==0.3.6
torch-sgld==0.1.0
comet-ml==3.37.1
betacal==1.1.0
POT==0.9.5